// src/components/interfaces/compositions/ProductCard/ProductCard.model.ts
export interface ProductCardModel {
  ProductImage: string;
  ProductTitle: string;
  ProductSubTitle?: string;
  ProductDescription?: string;
  ProductCategory?: string;
  ProductGenere?: string;
  ProductPrice: number;
  ProductForwardPrice?: number;
  ProductOldPrice: number;
  ProductPixDiscount?: number;
}